1.Implement PriQueue_Push with List_InsertDataDescently, make the implemention much simpler.

+++ testcase/test_priqueue.c
@@ -280,7 +280,7 @@ static void* ReadDataThread(void *p_param)
 
             if (g_writeDataFinished)
             {
-                LOG_A("Queue count:%d.\n", PriQueue_Count(queue));
+                LOG_A("Queue count:%d.\n", (int)PriQueue_Count(queue));
                 ShowMsgQueue(queue);
                 break;
             }
